backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
apscheduler==3.10.4
python-multipart==0.0.17
itsdangerous==2.2.0


